Refactor: Implement atomic design structure for ProtectedRoutes.jsx component
