+++ src/App.jsx
@@ -1,17 +1,9 @@
-import { Header } from "./components/orgamisms/Header.jsx";
-import { Home } from "./pages/Home.jsx";
-import { Footer } from "./components/Footer.jsx";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import { About } from "./pages/About.jsx";
-import { SignUp } from "./pages/SignUp.jsx";
-import { Login } from "./pages/Login.jsx";
-import { Products } from "./pages/Products.jsx";
-import { Design } from "./pages/Design.jsx";
-import { Contact } from "./pages/Contact.jsx";
 import { AuthenticationContext } from "./context/AuthenticationContext.jsx";
-import { UserLogged } from "./pages/UserLogged.jsx";
-import { NotFound } from "./pages/NotFound.jsx";
-import ProtectedRoute from "./components/ProtectedRoutes.jsx";
+import { Header } from "./components/orgamisms/Header.jsx";
+import { Footer } from "./components/Footer.jsx";
+import ProtectedRoute from "./components/routing/ProtectedRoutes.jsx";
+import { routes } from "./routes";
 
 function App() {
   return (
@@ -21,17 +13,18 @@ function App() {
           <Header />
 
           <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/about" element={<About />} />
-            <Route path="/signup" element={<SignUp />} />
-            <Route path="/login" element={<Login />} />
-            <Route path="/products" element={<Products />} />
-            <Route path="/contact" element={<Contact />} />
-            <Route path="/design" element={<Design />} />
-            <Route element={<ProtectedRoute redirectPath="/login" />}>
-              <Route path="/user" element={<UserLogged />} />
-            </Route>
-            <Route path="*" element={<NotFound />} />
+            {routes.map(({ path, element, isProtected }) =>
+              isProtected ? (
+                <Route
+                  key={path}
+                  element={<ProtectedRoute redirectPath="/login" />}
+                >
+                  <Route path={path} element={element} />
+                </Route>
+              ) : (
+                <Route key={path} path={path} element={element} />
+              ),
+            )}
           </Routes>
 
           <Footer />
